application/backend/secondchance_backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
babel==2.16.0
beautifulsoup4==4.12.3
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.4.0
chatbotAI==0.3.1.3
click==8.1.7
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.0.2
django-cors-headers==4.5.0
django-docs==0.3.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
imagesize==1.4.1
incremental==24.7.2
Jinja2==3.1.4
MarkupSafe==3.0.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==75.2.0
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.1.3
sphinxawesome-theme==5.3.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
zope.interface==7.1.0
